Some major refactoring. All behaviour should be as it was beforehand.
This refactoring moves out edge labels into their own datastructure, the transition. Previously an edge label was `NULL` for epsilon edges, or pointed to a string for a label. However we need to express three kinds of labels: epsilon transitions, human-readable labels, and single-character state transitions.

Using single-character labels for state transitions seemed cumbersome, so these three types have been split off as different transition types. This simplifies code generation (particularly for C's `switch` statement, and for on-the-fly execution), and relegates labels to exist for human-readable use only. This avoids confusing programatic constructs with human-readable conveniences.

The other purpose for this is to facilliate other transition types which may follow in the future - most notablly wildcards.