Added a parser for globs, which parses to the same actions shared by the simple regexp parser. It has its own lexer.
The grammar is trivial, and would not have been written using SID were it not a requirement for all front-ends since they are intended to share [source:/trunk/src/libre/parser.act parser.act]. So it's a little overkill on parsing for the payoff of centralising the NFA generation.

This is the first alternate front-end, and serves mainly to help get the linking and such sorted out.