Added provisions for multiple opaque values. These allow for state colourisation to combine multiple colours into one state, for example during DFA minimisation.
This is an odd change; the main rationalle is that when unioning two FSMs into the destination, I might want their colour to be the same. i.e. for them to produce the same token from a lexer, from two different regexps. So therefore the `void *` itself is used as a colour.

I'm not intending to mix colour `void *`s and non-colour `void *`s, because the pointer value itself is the colour, not an address.

So therefore we simply have an opaque list. For the CLI application these are strings; for libre, these are (presumbaly) some regexp struct. Hence `fsm_union()` takes a `void *` and appends it to the list for the entire incoming FSM. This does not detract from any existing opaque values states might have.