This addresses a bug found during minimization; for reversal with multiple end states, the new start state needs to be able to access that entire set. Previously this was done by introducing a new state, and transitioning via epsilons to each of the candidates in that set. However, this extra state prevented complete minimization in a few cases.
Instead, one of the states from that set is arbitarily nominated as the new start state, and linked to the other candidates in the set by epsilon transitions. This avoids needing to add any states. In effect, it is one giant state, all glued together.

Minimization bug spotted by Wulf.