Added interfaces to set colours per edge, including propogating them through on operations such as reversal, and NFA->DFA conversion.
I'm a lot happier with this; although I'm not bothering to be particularly tidy right now (because this is another experimental approach, which may again turn out to be yet another evolutionary dead end), it feels like this might actually be getting somewhere in the direction of splitting apart by colours. I hope.