Remove opaqued colour handling.
This was a bad idea, with a clumsy implementation due to its poor (i.e. nonexistant) theoretical basis. In particular, merging and splitting nodes causes difficulty.

This was origionaly motivated by the problem of keeping end states for unioned regexps distinct during minimization (such that the responsible regexp can be identified). However, flagging colours on states does not serve to prevent over-minimization from merging two regexps which ought to be seperate; colours would also need to be present on edges. That would be very complex, and I decided to look for other solutions instead.

If this returns, it will be in a different form. Most likely just the end states would need to carry an opaque `void *`, rather than arbitary states, and so that could be combined with the concept of indicating an "end" state. However, that would be a single item and not a set, so all this set-handing code can go.