Added boolean operations.
These will eventually take the place of `libre`'s equivalent operators.

This is also a start on reforming the API for sensibleness; the `fsm_merge()` interface is now made private. All binary operators (boolean and otherwise) are destructive to their operands, and return (the illusion of) a newly-created `struct fsm`.