Remove `struct re`; this was reduced to just contain a `struct fsm`, so that is now used directly.
This represents a mindset change; libre now exists soley to parse regular expression formats and to construct a finite automata. I think this makes sense, because there is no conceptual difference between a regular expression and a finite automata; the API should reflect this equivalence. Users are expected to be familiar with the theory here.

Execution remains in libre, mainly for convenience.