Remove return value for `lx->pop()`; this should never be doing things which can fail. (`realloc`ing down, for example, would be done in `lx->clear()` instead.)
Note `lx->pop()` may be requested to pop an already empty buffer, but only once (due to one-character lookahead for a DFA). This seems lax, but it is a side-effect of calls to `lx->push()` being optimised away when static analysis shows there are no accepting states which emit a token henceforth. That optimisation is lovely, and I don't mind having `lx->pop()` deal with one excess call of aftermath in exchange.