Switched from the `<lx.*.mk>` build system to the non-recursive Makefiles I wrote experimentally for bibfs.
These in turn derived from the `<lx.*.mk>` Makefiles, but as an experiment into doing things in a simpler way, and without invoking `${MAKE}` for each subdirectory. That experiment was a success, and I'm now sharing those between projects; libfsm is the first.

There is still work to be done for those; the (small) differences between bibfs' layout and here make that clear. That work can be done here, and merged upstream. For this commit, `share/mk` is verbatim.