Added syntax for arbitary expressions for patterns.
This extends the placeholder syntax for concatenation, and provides in addition operators for reverse, complement, subtraction and the usual regexp operators, as well as parentheses for precedence. This syntax is much the same as for arithmetic expressions in infix notation.

The regexp operators are not currently implemented; this commit is to get the syntax in place.