Bugfix for a false positive in the set of conflicts, as found for: {{{ /aa(aa)*/ -> $x; /aaa(aaa)*/ -> $y; }}}
Here `$y` appears in both a conflicting and non-conflicting DFA state. When converting to a DFA, the previous code for `carryopaque()` pointed to an existing `ast_mapping` struct from the NFA.

That's fine for the cases where there's no conflict. However, for conflicts, that shared `ast_mapping` meant the conflicting set (kept for the sake of error messages) would be shared between all DFA states where that token is used.

The fix here is to construct an `ast_mapping` to hold the conflicting set, rather than using an existing one.