Attach matching patterns to end states.
This commit adds the mechanics for identifying ambigious matches, which error out by default but may be permitted by way of a new `-a` flag. Patterns are printed where appropriate, if the `-p` flag is given.

These are stored in much the same manner as for lx, but simply as a set of patterns rather than an `ast_mapping` structure.