Cast to `(char)` for unsigned values.
These are produced from `strtoul()`, where a user specifies an unsigned value in the range `0`..`UCHAR_MAX` inclusive. Here the value is first confirmed to be within that range, and then converted to `char` (which of course may be unsigned). So I'm casting via `(unsigned char)` because that's how I think of the sequence of conversion, even though it's not neccessary.