Add a facility for "shoving" characters into the lexical stream when parsing regular expressions.
This is not something I want to provide for lx's generated lexers in general. I'm putting it here in `libre/parser.act` because the problem it solves is specific to libre.

The problem is, for entering the grammar at `::group-$dialect` (i.e. when invoking `re -g`), the lexer would usually have reached that point by making a zone transition. However, jumping to the grammar there, the lexer is left in its starting zone.

I didn't want to have the actions set the lexer's starting zone explicitly, because zones are numbered arbitrarily, and that's breaking abstraction. So I've solved this by having the parser "shove" characters into the input stream, both before and after the regular expression being parsed, just as if the user had typed them.

I prefer this approach because it keeps knowledge of this problem private to the parser actions, which I see as the parser's responsibility because it's the parser which allows multiple entry points into grammars in the first place.

The "shoving" mechanism is implemented by a state machine in `lgetc()`. It's not particularly nice. In particular there's another question of responsibility; I don't think the parser has any business knowing the spelling for a token, as I'm hardcoding here.

Future work might be to consider making it a little more general, to walk prefix and postfix strings rather than inserting single characters. Then those strings could come from the lexer's generated `lx_example()` interface, which would allow this to be written in terms of token names, rather than the spellings for those tokens.