Convert character-shoving state machine to walk strings, rather than single characters.
I had intended to write:
{{{
#define LX_EXAMPLE CAT(LX_PREFIX, _example)
lex_state->shove.pre  = LX_EXAMPLE(lx->z, @$OPENGROUP);
lex_state->shove.post = LX_EXAMPLE(lx->z, @$CLOSEGROUP); /* XXX: $closegroup does not exist in the current zone! */
}}}

Such that the parser actions can find adequate spellings for strings, and thus specify the prefix and postfix in terms of token names rather than strings.

However this has two problems: the current zone isn't set (so `lx->z` is `NULL`), and also the `$closegroup` token only exists inside the group's zone. So we'd need to be able to refer to the zone by name anyway, and the whole point of this mechanism was to avoid doing that.