Add "one-way" zones.
These are zones for which there is entry, but no exit.

One-way zones cannot be implicit.

Our motivating use-case here is to build a trie:
```
'a' {
	'b' {
	}
	'c' {
	}
}
```

This is quite unusual. It's a lexer you can only use once! and then for the next user-agent string, you'd `lx_init()` again.

The closest I can think of lexically would be e.g. HTTP, closing after one connection: there's one lexer zone for the headers, and then we make a one-way transition to another zone - and then a different zone for the body:

```
# model of http
/[a-z]+/ -> $str;
':' -> $sep;
"\r\n" -> $crlf;

"\r\n\r\n" {
	/./ -> $char;
}
```