set: improve encapsulation semantics of API
A set encapsulates some data. Prior to this commit, the set API returned the
encapsulating data storage instead of the encapsulated data. This meant that
consumers would frequently assume information about the set implementation,
making it difficult to change.

This change makes `struct state_set` opaque; we should now be able to replace
the set implementation with something else.