libfsm, priq: Fix shortest path calculation for fsm_example
This broke after the state list from an edge was no longer maintained in
insertion order, commit 69779c0c9a9bed0efb9d31766d9dc23e3cd15b0e. The
breakage only happened on FSMs with more than two states, and only if
the ordering of the edges was not insertion order.

The reason for the breakage was that fsm_shortest updated the priq in
Dijkstra without reordering the elements in the priq. This wasn't
problematic before the change, because destination states from an edge
being in insertion order gave some guarantees about how the path could
be reached.

A new function, `priq_update` is introduced to update the cost of a node
in a priq, maintaining the semantics of the structure. `fsm_shortest`
now calls this function when it finds a shorter path over some edge. And
we now have working examples again.