Set equality here was asymetric; this commit fixes that skew introduced during the set refactoring.
The arrangement here now checks for equality of the sizes of `a->edges` and `b->edges`, and the existing loop checks that all symbols match.

Curiously, this is almost identical to a bug I had in the same function, several years ago. The implementation here is different (as opposed to my usual `subsetof(a, b) && subsetof(b, a)` in order to keep the runtime cost down.