No group entry points.
This was a feature I rather liked; unfortunately it's based around the convenience of a bitmap representation. This no longer fits with pre-compiled classes (to be provided as arbitrary DFA).

This code is already more complicated than I'd like; keeping support for these features whilst changing the underlying representation would make it too unwieldy.

This commit also sees the "shove" mechanism removed, which I'm glad to see gone.