Remove the makefile rule for utf8_all.re; this is now always re-generated when re-building utf8_all.c.
This avoids the where utf8_all.re was not present for a fresh checkout, causing the generated .c to always be re-built - which of course it can't, since building that depends on `${RE}`.

Since the .re file is trivial to generate, my solution here is simply to fold those targets together.

Spotted by dho.