Introduce a prelude to the first state for lx's generated code.
This serves to distinguish an empty input stream (i.e. before any input is read) from `TOK_ERROR`.

I think this could also be expressed as matching within an FSM, but `TOK_EOF` is in general an artefact of the generated code anyway, and so I don't want to provide its behaviour in two places. In any case, the naieve approach to return `TOK_EOF` when in the start state would be incorrect, because that would be the state for `/x*y/` given input of `xxxx`, and that ought to be `TOK_ERROR`, not `TOK_EOF`.