Fixes small UB in lexers output by lx
In lx lexers that have a dynpush buffer, the code to increase the buffer
size would first realloc() and then determine the position in the new
buffer using pointer subtraction of the old pointer values.  This is
undefined behavior if realloc() returns a new value for the pointer,
because the old values will become indeterminate.  This patch moves the
subtraction to before the realloc() and updates the files generated by
lx