Fix bug caused by unintended sign extension.
The callback passed to `fsm_exec` returns int, to allow for EOF, but
the code paths that handled the result truncated the value to char and
then expanded it back to int, causing a potential sign extension bug.
In particular, characters >= 0x80 could be sign-extended to 0xffffff80
between `state = nextstate(state, c)` and `nextstate`'s call to
`fsm_hasedge`, which would lead to state match failures in an otherwise
correctly constructed FSM.

Avoiding the temporary truncation and keeping the edge token value
as an int fixes the sign extension.

(theft found this bug.)

Note: I haven't been able to run the tests, they seem to depend on sid,
and I'm not sure this hasn't broken something else. In particular,
it may break handling of epsilon, which is defined as the max unsigned
char + 1.