Rename `free_{SET}` functions to clear_{SET}` -- they only free contents.
free_trans and free_mappings don't free the collection, they only free
their contents and reset the set. Rename them to 'clear_*', and let
the explicit call to `set_free` in `fsm_determinise_freecache`
free the sets, since the trans set is reused after clearing it in
the for loop.