Add some theft-based property tests.
- literals
- PCRE regexes
- sets
- priqs
- NFAs

These depend on theft, and are not wired into the overall build system yet.
Not all of them are expected to pass, currently; some provide minimal
reproductions for current issues.

Adding a condition (aside from the initial length) that truncates
the state and edge counts when 0 DRAMATICALLY speeds up shrinking,
because it introduces more choice points to prune off unnecessary
elements.

- Add test: NFA edge creation order should not matter.

- Add minimal failure for fsm_minimise bug.

- Add a regression test for `fsm_trim` crashing without a start state.
  (This is addressed in a separate PR.)

- Trim before reversing; this is a workaround for a known issue.

- Don't call `fsm_exec` without determinising the NFA first. This
  applies before the operations, too. This is part of the API contract.

- Edit: Always set a start state.