Order subdirectories for innermost first.
This avoids a bug during `make clean` where subdirectories are attempted to be removed after their parents.

I rationalise that depending on order in my makefile here is not a kludge, because this is the ordering for variable content rather than for the makefile rules proper. Those should never depend on a particular ordering, since they are defined in terms of their dependencies.