Remove state equivalence folding from `fsm_minimise()`.
This was introduced to account for a situation whereby reversing would introduce an additional state, linking to multiple states by epsilons. The newly-added start state is not rechable by other means, and so transitions after the epsilon closure when converting NFA->DFA do not find those states equivalent.

However, merging them here is the wrong thing to do; my primary argument is that `fsm -rp | fsm -dp | fsm -rp | fsm -rp` ought to behave exactly the same as `fsm_minimise()` does.

So if this is reintroduced, it ought to be done inside `fsm_determinise()` instead.