Adds a few handy fuzzing/testing tools
gengraph.rb and gengraph_eps.rb are small Ruby scripts that generate
random NFAs and allow you to specify the number of nodes, the average
number of edges per node, and the maximum number of end states.
gengraph_eps.rb also allows epsilon transitions.

testintsub.sh is a bash driver script that generates a bunch of NFAs and
tests intersection and subtraction.

The output of intersection is compared with ~(~A | ~B).
The output of subtraction is compared with ~(~A | B).