Free arguments on success.
This is totally unneccessary here, and in some cases inconvenient to the caller. But I don't want to introduce an API change due to converting to `fsm_walk2()` here and leaking in the meantime is not desirable.

In the meantime this keeps the behaviour as it was, until I can think about it a bit more.