lx: Retain/restore column position around newlines in `*getc/*ungetc`.
This adds a field to `struct lx_pos`, `.saved_col`, which is used to
save and restore the column position when `getc`/`ungetc` handle '\n'.

If I'm understanding the test setup correctly, this change does not
cause any new test failures. I have a small standalone test program
which uses an lx-genearted lexer with `-b dyn -k str`, sinks various
known strings into it, and checks that the line/colum/byte positions are
as expected. I could add that to the PR, but I'm not sure how to
integrate a standalone .c test program into the build/test config.

This implementation wastes an unsigned (`.saved_col` is unused in
`lx->start`), which could be avoided, but would likely complicate
the other code generation.