subgraph.c: Fix use-after-free.
Running `(abc){3,5}d` under valgrind while working on the `sql`
dialect found a use-after-free here. It looks like the current
implementation can free `start` during the `mapping_free(mappings);`
at the end, so the resulting state should be grabbed first.

I'm not 100% sure all the details are correct, but this change
is sufficient to eliminate the valgrind warning.