lx: In `str` IO mode, don't count final '\0' toward end.byte pos.
This appears to be necessary to get accurate byte positioning -- see the
test code in the next commit. I'm not sure whether it should apply to
any of the other configurations, but this seems appropriate for `str` IO
mode since it's handling a series of '\0'-terminated strings, and the
'\0' at the end isn't actually part of the input.