tests/lx: Add lexer spec and harness for checking token positions.
This generates a small lexer (in `str` IO mode), gives it a series of
test strings, and checks that the tokens have the right types, string
buffer contents, and positions.

Note: this does not test anything else after the line comment; it
doesn't seem to exit the line comment zone, but I'm not sure if this
is a bug or API misuse.