First cut at a codegen IR.
Nodes here currently correspond to FSM states, although that need not be true for the future, especially when identifying parallel walks may be an option.

The intention is for decision-making about code generation to be made present explicitly in this IR, such that the code generation outputs things roughly as given. This way, I hope for various optimisations to be shared across multiple output languages, but also for the code output parts to be simpler.