Separate edge symbol escaping from character escaping.
It's rare that we need to print an epsilon symbol - currently only for dot output. That confuses the range checking for character values. Here I've separated out the `escputc` functions so that we can use C's type system to do that range checking, instead.

`escputc` is now passed `char` only (which has signedness per the implementation), and the newly-introduced `escpute` deals with the full `enum fsm_edge_type` range (which is unsigned), including special edges.