Convert lxpos tests to `lx -l dump` format.
To do this, I transcribed the C structs by hand, rather than trusting the existing `lx -l dump` output.

Some of the tests relied on continuation of lines between invocations; this is obviously not applicable for separate runs of the generated dump program, and so I duplicated their previous input where neccessary (and adapted line numbers to suit).