Fix failing tests -- adjust ALT productions in grammars with '|'.
ALT nodes are now collected into a regular linked list (every
right node is either another ALT node, or EMPTY), which is later
unrolled into a single array. (Eventually, this should happen in
the parser.)

This fixes a couple tests failing because due to an assert for
the ALT nodes' right node type.