Remove special root node case; add test for 'a'.
Previously, there was an edge case where the toplevel concat node
would pass along an incorrect linkage to its contents -- I think an
input that produced the failure was `(^ab|c)$` -- so there was a
check that always returned only top-down linkage for the root node
of the AST. This meant that input of a single character would be
incorrectly anchored.

Since the anchor nodes are retained and linked directly during
compilation now, that check no longer seems necessary -- the case
it tried to correct is already working. Remove the special case,
and add a regression test.