Fix false matches from consecutive backward epsilons at end of regex.
There was previously a false match for `a*b*` and `a*b+`, because
the epsilon repeating the `b` led back to the state _before_
matching the `a`, so it'd match "abab...".

This failure only occurred at the end of the string, because it
wasn't checking whether a state & epsilon guard needed to be
added before the very last node in a CONCAT. Now, check there,
and just skip the forward checks if the next node is NULL.

Add tests.