Changes mapping_set to use a hash set
Profiling showed that determinising a large graph could spend large portions of its time in `set_add`.  The culprit was the insertion sort used by `set_add`.  When adding a large number of items, this is quadratic.

mapping_set isn’t used in a way that requires it to be ordered, making it a perfect place to use a hash set.