sweep_states returns long instead of int
The number of states in the graph can be an `unsigned int`.  Because `sweep_states()` returns the number of trimmed states, if that number exceeds `INT_MAX`, undefined behavior will be invoked (most likely signed integer wrap-around).

Changed the return type to a `long` which will hopefully be wide enough to accomodate trimming more than `INT_MAX` states on most sane architectures.