Pass through src and dst FSMs separately for the `.carryopaque()` callback.
Here there was an assumption that `struct fsm_state` is self-contained enough to do what these implementations of the carryopaque callbacks needed to do. That's not true in general, however. And moreso the state will be passed as an index in the forthcoming adjacency list work. So here the FSM associated with the src state set and the dst state are passed explicitly.

The walk2 facility has no single state machine to pass, so it constructs a temporary FSM just for the sake of holding the one or two states it needs.