Remove support for identifying overlaps within character classes.
This was neat and I liked that it could be done. But it's also quite unneccessary, and the handling gets in the way internally. I'm removing it for lack of need, and for simplicity.

My rationale is that character classes are equivalent to `x|y|z` form, and we don't identify overlaps for those.