Move subtraction from the class tree up to expressions in general.
This leaves the class tree a little simpler, having no mutually recursive nodes, and sets the stage for further simplifications.

In this commit I also complete the implementation of subtraction for the SQL 2003 dialect. It's probably not quite conformant (for example I haven't checked all the situtions for where `^` is permitted), but that's not my focus right now.