Implement the AST invert node in terms of subtraction from `/./` instead of a special node.
I was attempting to avoid constructing intermediate FSM here, but haven't been able to avoid it. This implementation is more steps than before, but both are imperfect, and I prefer this for the simplicity elsewhere.

The abstraction works out here quite nicely; to the grammars it appears that we present an `<ast-make-invert>` action, and the details of how that node is implemented are kept out of (the already quite complex) parsing, *and* the AST tree compilation needn't handle inverting specially.