Adds a simple method to duplicate the subgraph in comp_iter_repeated
Method works because comp_iter only adds states, and states are
consecutive integers.  The subgraph is captured by saving the state
count before and after the AST_EXPR_REPEATED subnode is created.  This
captures all of the states that need to be copied.

Copying then involves duplicating the states of the subgraph and the
edges of the states within the subgraph.  This made easier because the
states within the new subgraph are ordered exactly the same as the
states within the old subgraph, so and edge in the original subgraph
(old_start .. old_end)

    old1 --[sym]--> old2

can be duplicated in the new subgraph (new_start .. new_end) by:

    new1 = new_start + (old1 - old_start);
    new2 = new_start + (old2 - old_start);

    new1 --[sym]--> new2