Disallow a literal nul byte for pcre.
From testing with both pcregrep(1) and perl's `//` syntax, this byte cannot appear in a regexp. The same value may be matched by spelling out the escape sequence `\x00`. So to account for this, I've left the `CHAR:` terminal extraction as it is (as arbitrary bytes), but disallow nul in the lexer for pcre specifically.