Undertake subtraction at AST time.
This addresses a flaw from introducing the subtraction node, where an empty result causes a nullable node, which affects the analysis for anchoring. This is resolved by performing the subtraction and replacing the node with `AST_EXPR_EMPTY` when necessary.

The subtraction itself is implemented by constructing FSM. This is expensive, but serves for the general case. It would be possible to implement faster subtraction for simpler cases before falling through to this.

See the attached test for an example.