Remove ordering and de-duplication by symbol for edge sets.
This is one of those cases that seems like going backwards in order to go forwards. Here edges are no longer inserted in order, and the various things dealing with edges are adjusted to cater for that.

My intention here is to allow for eventually storing bitmaps of symbols attached to each edge, instead.