Convert edge sets to a single state per edge struct, rather than a state set per edge struct.
This relies on duplication between multiple instances of `struct fsm_edge` within an edge set. Effectivly here, the concept of duplication of state items is merged with the concept of duplication of symbols; both are now implemented in the same place.

This simplfies a lot of the edge traversal code, but also allows for future work indexing edges by a (now single) state, rather than by a set of states.