Add .fsm syntax to allow declarations for states.
This provides a way to force numbering for states when parsing. Currently the labels are not numeric, and so states are numbered internally in their creation order. This is confusing when using .fsm as an intermediate format for serialisation in a pipeline, where one would expect the output to match the input.