Bugfix: Don't update i when array value was removed.
This bug has been dormant, but the new `fsm_trim` removes more
states and made it easier to encounter assert failures due to
stale edges sticking around.

A minimal input that provoked this was

```
0; 1;
0  ->  0 "A"; # start
0  ->  1 "X"; # e.g. "s"
0  ->  1 "s"; # e.g. "s"
1  ->  1 "X"; # e.g. "s"
1  ->  1 "s"; # e.g. "s"
start: 0;
end:   0;
```
which would delete the 0 -> 1 "X" edge, erroneously step over the
0 -> 1 "s" edge, and then hit an assert failure later when there
was a stale "s" edge to a state that no longer existed.