edgeset, stateset: Avoid redundant memmoves while compacting.
`state_set_compact` matches `edge_set_compact` and should be
correct, but it isn't currently tested -- `state_set_compact` is
only called when compacting a state's epsilon set, and trimming is
only done on DFAs now. It may be useful for merge or other
internal operations, otherwise it should be removed.