edgeset: Remove alloc pointer.
Don't store a pointer to the allocator on every single edge_set,
because there's only one function (`edge_set_free`) where the
alloc isn't already passed in, and it's going to be used in a
context where either the `struct fsm *` or its allocator callback
will be available.

This saves one pointer per state.