ir tests: Reorder expected output (workaround).
There is some degrees of freedom in the IR's json output, and
recent changes to `fsm_minimise` is causing output states to be
renumbered (in an equivalent way), which seems to be affecting the
JSON output. Since the current test is comparing the output contents
but doesn't understand which reorderings are functionally equivalent,
it's causing a few tests to fail.

While it'd be better to make the output deterministic, or to make
the test comparison more understanding, either is ultimately outside
the scope of this PR. After discussing on slack, @katef is good with
just updating the expected test output for now.