ir tests: Reorder expected output, again (workaround).
The updates to fsm_minimise have again reordered the JSON output,
breaking some of the ir tests. The real solution would be to
make output have a stable ordering, but that is outside the scope
of this PR.

They were also updated in ad1510e5dbafa527745a79f314682b36d51c9171.
Based on previous discussion, we decided to update the expected
json output and more directly address this later.