Do not assume long has 64 bits.
In hashset.c, siphash is called with a size of sizeof(unsigned long),
but siphash asserts that the size is either 8 or 16. Change h's type to
uint64_t instead. On platforms where long does not have 64 bits, this
will implicitly truncate the result in the return statement, which is
fine.

In minimise.c, use UINT64_C(1) instead of 1UL when a 64-bit constant is
needed.