Fix the handling of PCRE flags.
This removes the `AST_EXPR_FLAGS` node type. This node type was used to
indicate that the following nodes should be interpreted according to its
flags, but "the following" was implemented as "the following children of
its parent", rather than "the following elements in the original syntax
until a closing parenthesis". "The following" was affected by internal
re-ordering of nodes in alternatives, and "until a closing parenthesis"
could not be determined anyway as non-capturing groups do not get their
own nodes.

This commit removes the `AST_EXPR_FLAGS` node entirely and instead
modifies the parser to pass the current `re_flags` to all AST
construction functions and stores it in a new field in `struct
ast_expr`. The compilation in `src/libre/ast_compile.c` is updated to
look at this field to find whether RE_ICASE is enabled (the only flag
currently implemented).

This commit updates the printing of ASTs as trees, dotfiles and PCREs to
reconstruct the flags. As they are no longer represented as nodes, they
do not get printed as nodes.