Don't subtract the exit pattern from bindings within a zone.
This was inappropriate; we have no way of knowing how the binding would be used. The reported case was of the form:
```
/a/ .. /b/ {
	v = 'b';
	'a' v -> $t;
}
```

where it would never be possible for `$t` to share a common subset with the exit pattern, because of the concatenation.

The exit pattern is still subtracted from the patterns which map to tokens.