Changes AST parsing to use a "pooled" allocator
AST nodes are allocated out of a "pool" which is basically a bump
allocator.  If an error is encountered in parsing, the entire pool is
released.

When the AST is freed, the entire pool is also released.

This version relies on a global pool variable that's later associated
with the `struct ast`, which makes parsing not thread-safe.