ast: allow for node cloning
When constructing an AST directly, as opposed to parsing a regular
expression, it may be desirable to clone nodes. This commit adds an
ast_expr_clone function for this purpose. It is not yet used anywhere.

It returns 1 if cloning succeeded, or 0 if cloning failed, and modifies
the passed in expression pointer directly, rather than returning it by
value, so that callers can easily tell whether cloning succeeded.
Cloning a null pointer is valid and leaves the node as a null pointer.
If this had been returned by value, callers would have no easy way of
determining whether the clone operation succeeded.