Fix two issues in PCRE printing.
Fix unused local variable localflags: the idea of PR #249 was to use
this variable in the call to pp_iter, but this was accidentally still
using re_flags. This had no effect because we never need to group an
operand of |, as | has the lowest precedence of all operators, but we
may want to group operands of | in the future for simpler flag printing.

Fix printing of changed flags to ignore a change in global flags:
RE_ANCHORED applies to whole regexps only, not individual
subexpressions, and cannot be turned on or off. This cannot occur when
constructing ASTs from regexps, as the RE_ANCHORED flag will be the same
for all AST nodes, but can happen when constructing ASTs from FSMs when
additionally using the -b option (which should have no effect).