tests/capture/capture4: Fix failure caused by carryopaque bug.
The carryopaque callback was setting `src_set[i]`'s opaque to `NULL`,
but in this case that was incorrect -- we actually expected it to set
that end state's opaque (and, thus, information about which of the
unioned DFAs match) on multiple destination end states.

This corresponded to `/ab*c/` and `/abc/` having an overlapping end
state for `"abc"`, indicating that input matches both.