Switch end-ID API to scalable approach; satify property tests; cleanup.
Switch from a fairly naive array-based approach to using a dynamic
hash table, keyed by the state ID. The endid array still has O(n)
insertion (checking for duplicates), but if necessary that could be
changed to a bitset or other approach in a separate pass, when we
have good real-world test data with more than a single-digit number
of end IDs associated with any state.

Add `fsm_endid_iter_state` to the API, because many use cases will
only need to iterate over a single state's end IDs.

Rework the end ID management in `fsm_glushkovise` -- that code is
the only caller where the source and destination IDs for the opaque
and endid operations are the same. That interacts badly with the
hash table implementation, because adding to it while iterating
over it can trigger a resize and then corruption during iteration.
Instead, add separate passes to collect info to add, and then add
it afterward. This may have also been the source of use-after-free
bugs I ran into in my capture ID test code a while ago.

Switch some of the functions to returning enums, to better distinguish
between allocation error, adding an element, or skipping because the
element was already present. Some renaming and other misc. cleanup.